==> 3_Preparacion/01 Crear dataset 10k sin fotos corruptas (por limpiar).ipynb <==
**Función:** crea listado de fotos para el sub dataset 10k (por limpiar). No incluye ficheros corruptos.

**Requiere:**
- ../datasets/photos (con todas las fotos del dataset)
- ../datasets/photos_train.json (fichero json con labels de todas las fotos)

**Crea:**
- ../datasets/corrupt_ids.json
- ../datasets/df_10k.json

import pandas as pd
import json
import os
from PIL import Image

ejecutadoEnGoogleDrive = False # True si se ejecuta en Google Drive

if ejecutadoEnGoogleDrive:
    from google.colab import drive
    drive.mount('/content/drive')
    PATH = '/content/drive/clasificacion_imagenes'
else:
    PATH = '../'

# Detección de fotos corruptas

# Loop por todas las fotos para ver IDs que no abren. 
folder = PATH + 'datasets/photos'
item_list = os.listdir(folder)
ok_ids = []
corrupt_ids = []
for i in item_list:
    try:
        im = Image.open(folder + "/" + i)
    except:
        splitter = i.split(".")[0]
        corrupt_ids.append(splitter)

# Guardar los ids de fotos corruptas
with open(PATH + 'datasets/corrupt_ids.json', 'w') as corrupt_json:
    json.dump(corrupt_ids, corrupt_json)

print(corrupt_ids)

print(len(item_list) - len(corrupt_ids))

# Eliminación de corruptos

# Read JSON
json_train_path = PATH + 'datasets/photos_train.json'
df = pd.read_json(json_train_path, lines=True)
df.head()

df.info()

df_clean = df[df['photo_id'].isin(corrupt_ids) == False]
df_clean.info()


print(df.shape[0] - df_clean.shape[0])

# Creación del sub dataset 10k (por limpiar)
(misma proporción que el dataset completo)

df.label.value_counts() / df.shape[0]

# Create a dataframe. No usar clean_df para no contar con archivos corruptos. 
def create_df(num_samples, dataframe):
    labels = list(df.label.value_counts().reset_index()['index'])
    proportion = list(df.label.value_counts() / df.shape[0])   
    dict_samples = {}
    for (i, j) in zip(labels, proportion):
        dict_samples[i] = int(round(j * num_samples, 0))
    df_short = pd.DataFrame()
    for i in dict_samples.keys():
        df_temp = dataframe[dataframe['label'] == i].sample(dict_samples[i])
        df_short = pd.concat([df_short, df_temp], axis=0)
    return df_short
df_reduced = create_df(10000, df_clean)
df_reduced.label.value_counts()

# Check archivos únicos
df_reduced.photo_id.nunique()

df_clean_json = df_reduced.to_json(PATH + 'datasets/df_10k.json', orient = 'records')

# mini EDA del 10k

# Ejemplo numero de fotos por categoría
labels = list(df.label.value_counts().reset_index()['index'])
proportion = list(df.label.value_counts() / df.shape[0])
fotos = 10000
cummulative = []
for i in proportion:
    cummulative.append(round(i * fotos, 0))
print(dict(zip(labels, cummulative)))
print(sum(cummulative))
