==> 3_Preparacion/02 Identificar duplicados del 10k (por limpiar).ipynb <==
!pip install difPy

import os
import json
import shutil
import pandas as pd

from difPy import dif


json_path = '/Users/tonivila/Downloads/df_10k.json'

df = pd.read_json(json_path)
df.head()

def duplicates_finder(label, dir_ori, dir_dest, save_dir):
    pic_ids = list(df[df['label'] == label].photo_id)

    for i in pic_ids:
        try:
            shutil.copy(dir_ori + '/' + i + '.jpg', dir_dest)

        except:
            pass

    search = dif(dir_dest, show_progress=True)
    results = search.result

    with open(save_dir + f'{label}_duplicates.json', 'w') as sv:
        json.dump(results, sv)
    
    return results




origin = '/Users/tonivila/Downloads/imagenes_negocios'
destination = '/Users/tonivila/Downloads/menu'
save = '/Users/tonivila/Downloads/duplicates_json'


duplicates_finder('menu', 
                    origin,
                    destination,
                    save)
