==> 3_Preparacion/03 Visualizar duplicados 10k (por limpiar).ipynb <==
**Función:** visualiza duplicados de la label 'drink' del dataset 10k (por limpiar)

**Requiere:**
- ../datasets/drinks_duplicates.json

**Crea:** (no crea ficheros)

import os
import time
import json
from time import sleep
import pandas as pd
import matplotlib.pyplot as plt
import pickl
from PIL import Image
from IPython.display import clear_output

CATEGORIA = 'drink' # label a revisar

ejecutadoEnGoogleDrive = False # True si se ejecuta en Google Drive

if ejecutadoEnGoogleDrive:
    from google.colab import drive
    drive.mount('/content/drive')
    PATH = '/content/drive/clasificacion_imagenes'
else:
    PATH = '../'

drink_duplicated = PATH + 'datasets/drinks_duplicates.json'

df_duplicated = pd.read_json(drink_duplicated, orient='index')
df_duplicated['location_clean'] = df_duplicated['location'].apply(lambda x: x.split('/')[-1])
df_duplicated['duplicates_clean'] = df_duplicated['duplicates'].apply(lambda x: str(x).split('/')[-1][:-2])


df = df_duplicated[['filename', 'duplicates_clean']].reset_index(drop=True)

df.head()

df.filename.nunique() - df.duplicates_clean.nunique() 

df.duplicates_clean.value_counts()

df[df['duplicates_clean'] == 'PUn1BLjRf1Go5wfCltqk1w.jpg']

df[df['duplicates_clean'] == '8aFn30vJyWkpsGYPSTSFDg.jpg']

example = list(df[df['duplicates_clean'] == 'PUn1BLjRf1Go5wfCltqk1w.jpg'].filename)
print(example)

folder = PATH + 'datasets/photos'

fig = plt.figure(figsize=(10, 20))
position = 1
for i in example:
    fig.add_subplot(5, 3, position)
    im = Image.open(folder + '/' + i)
    plt.title(i)
    plt.imshow(im)
    position += 1

list_duplicated = list(df_duplicated.filename)
print(len(list_duplicated))

fig = plt.figure(figsize=(10, 120))
position = 1
for i in list_duplicated:
    fig.add_subplot(35, 3, position)
    im = Image.open(folder + '/' + i)
    plt.title(i)
    plt.imshow(im)
    position += 1
