==> 3_Preparacion/04 Eliminar duplicados del 10k.ipynb <==
**Función:** crea sub dataset 10k (limpio)

**Requiere:**
- ../datasets/df_10k.json

**Crea:**
- ../datasets/df_10k.json

**ATENCIÓN:** en esta versión el fichero de salida tienen el mismo nombre que el fichero de entrada. Se recomienda crear una cópia del fichero 'datasets/df_10k.json' antes de ejecutar este notebook.

import pandas as pd
import numpy as np

ejecutadoEnGoogleDrive = False # True si se ejecuta en Google Drive

if ejecutadoEnGoogleDrive:
    from google.colab import drive
    drive.mount('/content/drive')
    PATH = '/content/drive/clasificacion_imagenes'
else:
    PATH = '../'

df = pd.read_json(PATH + "datasets/df_10k.json")
df.head()

def load_df(label):
  filename = PATH + 'datasets/{}_duplicates.json'.format(label)
  print('Processing {}'.format(filename))
  df_duplicated = pd.read_json(filename, orient='index')
  df_duplicated['location_clean'] = df_duplicated['location'].apply(lambda x: x.split('/')[-1])
  df_duplicated['duplicates_clean'] = df_duplicated['duplicates'].apply(lambda x: str(x).split('/')[-1][:-2])
  df_duplicated = df_duplicated[['filename', 'duplicates_clean']].reset_index(drop=True)
  return df_duplicated

duplicated_labels = ['food', 'inside', 'outside', 'drink']
d1 = df.copy().set_index('photo_id')
for label in duplicated_labels:
  d2 = load_df(label)
  d2['photo_id'] = d2.filename.apply(lambda x: x[:-4])
  d2 = d2.set_index('photo_id')
  d1 = d1.join(d2, how='left')
  d1 = d1.loc[d1.duplicates_clean.isna()].drop(columns=['filename','duplicates_clean'])
  print('After processing {} now remain {} photos'.format(label, len(d1)))
 

d1 = d1.reset_index()
d1.head()

json_ = d1.to_json(PATH + '/datasets/df_10k.json', orient = 'records')
