==> 3_Preparacion/06 Crear sub dataset 10k limpio.ipynb <==
**Función:** crear un sub dataset 10k limpio combinando los 5 ficheros de fotos reclasificadas

**Requiere:**
- ../datasets/food.json
- ../datasets/inside.json
- ../datasets/outside.json
- ../datasets/drink.json
- ../datasets/menu.json

**Crea:**
- ..datasets/df_clean_withKO.json
- ../datasets/df_clean.json

import pandas as pd
import json

ejecutadoEnGoogleDrive = False # True si se ejecuta en Google Drive

if ejecutadoEnGoogleDrive:
    from google.colab import drive
    drive.mount('/content/drive')
    PATH = '/content/drive/clasificacion_imagenes'
else:
    PATH = '../'

df_food = pd.read_json(PATH + 'datasets/food.json')
df_inside = pd.read_json(PATH + 'datasets/inside.json')
df_outside = pd.read_json(PATH + 'datasets/outside.json')
df_drink = pd.read_json(PATH + 'datasets/drink.json')
df_menu = pd.read_json(PATH + 'datasets/menu.json')
df_revisado = pd.concat([df_food, df_inside, df_outside, df_drink, df_menu], axis=0)
df_revisado.reset_index(inplace=True, drop=True)
df_revisado.tail()

df_revisado = df_revisado.to_json(PATH + 'datasets/df_clean_withKO.json', orient='records')

df_revisado.reclass_label.value_counts()

df_final = df_revisado[df_revisado['reclass_label'].isin(['KO']) == False]
df_final.reset_index(inplace=True, drop=True)
df_final.tail()

df_final.reclass_label.value_counts()

df_final = df_final[['photo_id', 'reclass_label']]

df_final

df_final = df_final.to_json(PATH + 'datasets/df_clean.json', orient='records')
