==> 3_Preparacion/09 Crear sub dataset 10k balanceado limpio.ipynb <==
**Función:** crear un sub dataset 10k balanceado limpio combinando 5 ficheros de fotos reclasificadas

**Requiere:**
- ../datasets/food_rev.json
- ../datasets/inside_rev.json
- ../datasets/outside_rev.json
- ../datasets/drink_rev.json
- ../datasets/menu_rev.json
- ../datasets/final_rev_menu.json
- ../datasets/final_rev_outside.json
- ../datasets/final_rev_drink.json

**Crea:**
- ../datasets/df_final_bal.pkl

import pandas as pd
import pickle

ejecutadoEnGoogleDrive = False # True si se ejecuta en Google Drive

if ejecutadoEnGoogleDrive:
    from google.colab import drive
    drive.mount('/content/drive')
    PATH = '/content/drive/clasificacion_imagenes'
else:
    PATH = '../'

food_1 = 'food_rev.json'
inside_1 = 'inside_rev.json'
outside_1 = 'outside_rev.json'
drink_1 = 'drink_rev.json'
menu_1 = 'menu_rev.json'
# fotos adicionales que no estaban en 10k (para incrementar labels menos representadas)
menu_2 = 'final_rev_menu.json'
outside_2 = 'final_rev_outside.json'
drink_2 = 'final_rev_drink.json'

df_food_1 = pd.read_json(PATH + 'datasets/' + food_1)
df_inside_1 = pd.read_json(PATH + 'datasets/'+ inside_1)
df_outside_1 = pd.read_json(PATH + 'datasets/'+ outside_1)
df_drink_1 = pd.read_json(PATH + 'datasets/'+ drink_1)
df_menu_1 = pd.read_json(PATH + 'datasets/'+ menu_1)
df_menu_1.head(2)

df_outside_2 = pd.read_json(PATH + 'datasets/' outside_2, orient='index').reset_index(drop=False).rename(columns=
                                                                                               {'index': 'photo_id',
                                                                                                0: 'original_label',
                                                                                                1: 'reclass_label',
                                                                                                2: 'status'})
df_drink_2 = pd.read_json(PATH + 'datasets/' + drink_2, orient='index').reset_index(drop=False).rename(columns=
                                                                                           {'index': 'photo_id',
                                                                                                0: 'original_label',
                                                                                                1: 'reclass_label',
                                                                                                2: 'status'})
df_menu_2 = pd.read_json(PATH + 'datasets/' + menu_2, orient='index').reset_index(drop=False).rename(columns=
                                                                                        {'index': 'photo_id',
                                                                                                0: 'original_label',
                                                                                                1: 'reclass_label',
                                                                                                2: 'status'})
df_menu_2.head(2)

df_menu_1.columns

df_menu_2.columns

df_revisado = pd.concat([df_food_1,
                         df_inside_1,
                         df_outside_1,
                         df_drink_1,
                         df_menu_1,
                         df_outside_2,
                         df_drink_2,
                         df_menu_2], axis=0)
df_revisado.head(2)

# df_revisado = df_revisado.to_json('/content/drive/Shareddrives/The Valley - TFM/Ejercicio clasificación imagenes/Ejercicio 2 - Balanced/df_clean_withKO.json', orient='records')

df_revisado.reclass_label.value_counts()

df_limpias = df_revisado[df_revisado.reclass_label != 'KO']

df_limpias.tail()

df_limpias.reclass_label.value_counts()

df_limpias = df_limpias[['photo_id', 'reclass_label']]

df_limpias

categorias = ['food', 'inside', 'drink', 'outside']

balanced = df_limpias.loc[df_limpias.reclass_label == 'menu']
num_photos = len(balanced)
for cat in categorias:
    balanced = pd.concat([balanced, df_limpias.loc[df_limpias.reclass_label == cat].sample(n=num_photos,
                                                                                           replace=False)])
balanced.reclass_label.value_counts()

pickle.dump(balanced, open(PATH + 'datasets/df_final_bal.pkl','wb'))
