==> 3_Preparacion/10 crear train y test de 10k balanceado en ndarray de fotos 224x224.ipynb <==
**Función:** crear un sub dataset 10k limpio balanceado, en formato ndarray.

**Requiere:**
- ../datasets/df_final_bal.pkl

**Crea:**
- ../datasets/X_train.pkl
- ../datasets/X_test.pkl
- ../datasets/y_train.pkl
- ../datasets/y_test.pkl

ejecutadoEnGoogleDrive = False # True si se ejecuta en Google Drive

if ejecutadoEnGoogleDrive:
    from google.colab import drive
    drive.mount('/content/drive')
    PATH = '/content/drive/clasificacion_imagenes'
else:
    PATH = '../'

import pandas as pd
import pickle
import numpy as np
import math
import torchvision.transforms as T
from sklearn.preprocessing import OneHotEncoder
from tqdm.notebook import trange, tqdm
from PIL import Image

photo_path = PATH + 'dataset/photos'
img_size = 224

# Create Train Test DFs. 

df = pickle.load(open(PATH + 'datasets/df_final_bal.pkl','rb')).rename(columns={'reclass_label': 'label'})
df.head(2)

df.label.value_counts()

round(df.label.value_counts() * 0.8, 0)

def create_df(num_samples, dataframe):
    labels = list(df.label.value_counts().reset_index()['index'])
    proportion = list(df.label.value_counts() / df.shape[0])
    
    dict_samples = {}
    
    for (i, j) in zip(labels, proportion):
        dict_samples[i] = int(round(j * num_samples, 0))

    df_short = pd.DataFrame()

    for i in dict_samples.keys():
        df_temp = dataframe[dataframe['label'] == i].sample(dict_samples[i])
        df_short = pd.concat([df_short, df_temp], axis=0)

    return df_short


train_num = round(df.shape[0] * 0.8, 0)
df_train = create_df(train_num, df)

df_train.reset_index(drop=True, inplace=True)
df_train.tail()

compared = df.merge(df_train.drop_duplicates(), on='photo_id', how='left', indicator=True)
compared.head(2)

df_test = compared.loc[compared._merge == 'left_only'].drop(columns=['_merge', 'label_y']).rename(columns={'label_x':'label'})
df_test.head(2)

print(f'El tamaño de Train es de {df_train.shape[0]} y el de Test es de {df_test.shape[0]}')

# Reescalar

def resize_pictures(dataframe, img_path, size_rescaled):
    df_temp = dataframe.copy()
    df_temp['image'] = ''
    
    for im_pos in tqdm(range(len(df_temp)), desc='Photos cropped', miniters=len(df_temp)//100):
        
        with Image.open('{}/{}.jpg'.format(img_path, df_temp.iloc[im_pos].photo_id)) as image:

            width = image.size[0]
            height = image.size[1]

            # redimensionar
            if (height < width): # imagen estrecha
                height = int(size_rescaled)
                height = math.floor(size_rescaled * width/height)
            else: # imagen ancha
                width = math.floor(size_rescaled * height/width)
                height = int(size_rescaled)

            resized = T.Resize((height, width))(image)
            cropped = T.CenterCrop(size=int(size_rescaled))(resized)
            df_temp.iloc[im_pos, 2] = cropped

    return df_temp

# Create new train DF
df2_train = resize_pictures(df_train, photo_path, img_size)

# Create new Test DF
df2_test = resize_pictures(df_test, photo_path, img_size)

np.array(df2_train.iloc[1, 2]).shape

display(df2_train.head(2))
display(df2_test.head(2))

np.array(df2_train.iloc[1, 2].getdata())

series_X_train = df2_train.image.map(lambda img: np.array(img.getdata()).reshape(img_size, img_size, 3))
X_train = np.stack(series_X_train.to_numpy(copy=True))
X_train.shape

X_train = X_train.astype('float32')
X_train /= 255

series_X_test = df2_test.image.map(lambda img: np.array(img.getdata()).reshape(img_size, img_size, 3))
X_test = np.stack(series_X_test.to_numpy(copy=True))
X_test.shape

X_test = X_test.astype('float32')
X_test /= 255

pickle.dump(X_test, open(output_path + 'X_test.pkl','wb'))

categories = [['food', 'inside', 'outside', 'drink', 'menu']]
enc = OneHotEncoder(categories=categories, drop=None)

y_train = enc.fit_transform(df_train[['label']]).astype('float32').toarray()

y_test = enc.fit_transform(df_test[['label']]).astype('float32').toarray()

type(y_train)

img_rows, img_cols, img_ch = 224, 224, 3
num_labels = 5

# test de los datos de entrada
if   (X_train.shape[1] != img_rows) \
  or (X_train.shape[2] != img_cols)\
  or (X_train.shape[3] != img_ch)\
  or (y_train.shape[1] != num_labels)\
  or (X_train.dtype != 'float32')\
  or (X_test.shape[1] != img_rows) \
  or (X_test.shape[2] != img_cols)\
  or (X_test.shape[3] != img_ch)\
  or (y_test.shape[1] != num_labels)\
  or (X_test.dtype != 'float32')\
  or (y_train.dtype != 'float32')\
  or (y_test.dtype != 'float32')\
  or (X_train.shape[0] != y_train.shape[0])\
  or (X_test.shape[0] != y_test.shape[0]\
  or type(X_train) != np.ndarray\
  or type(X_test) != np.ndarray\
  or type(y_train) != np.ndarray\
  or type(y_test) != np.ndarray):
  raise Exception("El dataset no tiene las dimensiones esperadas!")
else:
  print('El dataset tiene las dimensiones esperadas')

pickle.dump(X_train, open(PATH + 'datasets/X_train.pkl','wb'))

pickle.dump(X_test, open(PATH + 'datasets/X_test.pkl','wb'))

pickle.dump(y_train, open(PATH + 'datasets/y_train.pkl','wb'))

pickle.dump(y_test, open(PATH + 'datasets/y_test.pkl','wb'))
